Moved Main to core
